public

+++ src/promedio/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
         return multiplicacion;
     }
     
-    public double division (int a, int b) {
+    public double division (double a, double b) {
         double division = a / b;
         return division;
     }
